requirements.txt
pandas==2.2.2
polars==0.20.31
pyarrow==15.0.2
numpy==1.26.4
scikit-learn==1.5.0
lightgbm==4.3.0
xgboost==2.0.3
imbalanced-learn==0.12.3
optuna==3.6.1
shap==0.45.1
networkx==3.3
haversine==2.8.1
tqdm==4.66.4
loguru==0.7.2
pyyaml==6.0.1
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0
pytest==8.2.2
pytest-cov==5.0.0
